+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD
 
+* Move mutator to request package; ignore missing mutators in client
 * Rename various New functions in mongo packages to NewStore for consistency
 * Add new fields to cgm settings data type
 * Add new fields to pump settings data type